% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross_1year}
\alias{cross_1year}
\title{Master's students' employment condition}
\format{\code{data.frame} with 3282 observations of 7 variables:
\itemize{
\item \code{employment_sit} - employment situation, a
\code{factor} with three levels: \emph{Employed}, \emph{Unemployed}, \emph{Trainee}.
\item \code{gender} - gender, a \code{factor} with two levels: \emph{Male}, \emph{Female}.
\item \code{finalgrade} - final grade degree, a \code{factor} with three levels:
\emph{Low}, \emph{Average}, \emph{High}.
\item \code{duration} - duration of studies, a \code{factor} with three levels:
\emph{Short}, \emph{Average}, \emph{Long}.
\item \code{social_class} - social class, a \code{factor} with five levels:
\emph{Working class}, \emph{White-collar workers}, \emph{Lower middle class},
\emph{Upper middle class}, \emph{Unclassified}.
\item \code{irregularity} - irregularity indicator of student's studies,
a \code{factor} with three levels: \emph{Low}, \emph{Average}, \emph{High}.
\item \code{hsscore} - high school final score, a \code{numeric} between 60 and
100.
}}
\description{
\code{data.frame} with 3282 cross-sectional observations of 7 variables about
employment condition of master's students one year after graduation. Data
refer to students graduated at the University of Trento (Italy) between 2009
and 2013.
}
\references{
Santi F., M. M. Dickson, G. Espa (2019)
"A graphical tool for interpreting regression coefficients of trinomial logit
models", \emph{The American Statistician}, \strong{73}(2), pp. 200-207.
\doi{10.1080/00031305.2018.1442368}
}
\keyword{data}
