% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_vdelta}
\alias{get_vdelta}
\title{It computes the vector of covariate change}
\usage{
get_vdelta(delta, model)
}
\arguments{
\item{delta}{see \code{\link{field3logit}}.}

\item{model}{object of class \code{model3logit}returned by \code{\link[=extract3logit]{extract3logit()}}.}
}
\value{
\code{numeric} vector of covariate change \eqn{\Delta\in\textbf{R}^k}.
}
\description{
Given the argument \code{delta} passed to \code{\link{field3logit}} either as
a \code{numeric} vector, a \code{character} or an \code{expression} (see
\code{\link{field3logit}}), \code{get_vdelta} returns the \code{numeric} vector of
covariate change \eqn{\Delta\in\textbf{R}^k}.
}
\keyword{internal}
