% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\alias{labels<-}
\title{Set the labels of a \code{field3logit} or a \code{multifield3logit} object}
\usage{
labels(object) <- value
}
\arguments{
\item{object}{a \code{field3logit} or a \code{multifield3logit} object.}

\item{value}{a character with the new label (or labels in case of a
\code{multifield3logit} object).}
}
\value{
Object of same class of argument \code{object} (either \code{field3logit} or
\code{multifield3logit}).
}
\description{
It sets the labels of an existing \code{field3logit} or a \code{multifield3logit}
object.
}
