% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field3logit.R
\name{simplify_field3logit}
\alias{simplify_field3logit}
\alias{is_simplified_field3logit}
\title{\code{field3logit} simplification function and test}
\usage{
simplify_field3logit(x)

is_simplified_field3logit(x)
}
\arguments{
\item{x}{an object of class \code{field3logit}.}
}
\description{
Given an object of class \code{field3logit}, \code{\link[=simplify_field3logit]{simplify_field3logit()}} returns
it in the simplified form. Function \code{\link[=is_simplified_field3logit]{is_simplified_field3logit()}} tests
whether an object of class \code{field3logit} is in the simplified form.
}
\keyword{internal}
