% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_common.R
\name{v2vedge}
\alias{v2vedge}
\title{Computes the coordinates of a vertex on the edge}
\usage{
v2vedge(v, edge)
}
\arguments{
\item{v}{\code{numeric} vector of ternary coordinates.}

\item{edge}{\code{numeric} value of the edge.}
}
\value{
\code{numeric} vector of ternary coordinates of the vertex on the edge.
}
\description{
Given the coordinates of one of vertices, \code{\link[=v2vedge]{v2vedge()}} computes the
coordinates of the vertex on the edge.
}
\keyword{internal}
