% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{versor}
\alias{versor}
\title{Computes the versor}
\usage{
versor(k, n, simplify = TRUE)
}
\arguments{
\item{k}{order of the versor.}

\item{n}{dimension of the space.}

\item{simplify}{if \code{TRUE}, a \code{numeric} is returned, otherwise
\code{versor} returns a column matrix.}
}
\description{
Given an \code{n}-dimensional space, the \code{k}-th versor is returned.
}
\keyword{internal}
