% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_region.R
\name{add_confregions_field3logit}
\alias{add_confregions_field3logit}
\title{It adds the confidence regions to a "field3logit" object}
\usage{
add_confregions_field3logit(x, conf = 0.95, npoints = 100)
}
\arguments{
\item{x}{an object of class \code{field3logit}.}

\item{conf}{confidence level of the region.}

\item{npoints}{number of points of the border.}
}
\value{
Object of class \code{field3logit} with updated confidence regions.
}
\description{
Given the confidence level, it computes the confidence regions
the effects for each arrow of the \code{field3logit} object.
}
\examples{
data(cross_1year)

mod0 <- nnet::multinom(employment_sit ~ gender + finalgrade, data = cross_1year)
field0 <- field3logit(mod0, 'genderFemale')
plot3logit:::add_confregions_field3logit(field0)

}
\keyword{internal}
