% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross_1year}
\alias{cross_1year}
\title{Master's students' employment condition}
\format{
\code{data.frame} with 3282 observations of 7 variables:
\describe{
\item{employment_sit:}{employment situation, a
factor with three levels: \emph{Employed}, \emph{Unemployed}, \emph{Trainee}.}
\item{gender:}{gender, a factor with two levels: \emph{Male}, \emph{Female}.}
\item{finalgrade:}{final grade degree, a factor with three levels:
\emph{Low}, \emph{Average}, \emph{High}.}
\item{duration:}{duration of studies, a factor with three levels:
\emph{Short}, \emph{Average}, \emph{Long}.}
\item{social_class:}{social class, a factor with five levels:
\emph{Working class}, \emph{White-collar workers}, \emph{Lower middle class},
\emph{Upper middle class}, \emph{Unclassified}.}
\item{irregularity:}{irregularity indicator of student's studies,
a factor with three levels: \emph{Low}, \emph{Average}, \emph{High}.}
\item{hsscore:}{high school final score, a numeric between 60 and
100.}
}
}
\description{
\code{data.frame} with 3282 cross-sectional observations of 7 variables about
employment condition of master's students one year after graduation. Data
are used in \insertCite{santi2019;textual}{plot3logit} and refer to students
graduated at the University of Trento (Italy) between 2009 and 2013.
}
\references{
\insertAllCited{}
}
\keyword{data}
