% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl2matrix}
\alias{tbl2matrix}
\title{Convert a tibble to a matrix}
\usage{
tbl2matrix(x, .rownames = NULL)
}
\arguments{
\item{n}{dimension of the space.}

\item{k}{order of the versor.}

\item{simplify}{if \code{TRUE}, a \code{numeric} is returned, otherwise
\code{versor} returns a column matrix.}
}
\description{
Given an \code{n}-dimensional space, the \code{k}-th versor is returned.
}
\examples{
plot3logit:::versor(1, 3)
plot3logit:::versor(3, 6)
plot3logit:::versor(3, 6, FALSE)

}
\keyword{internal}
