% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf-region.R
\name{vcovB2vcovDeltaB}
\alias{vcovB2vcovDeltaB}
\title{Covariance matrix of covariate change}
\usage{
vcovB2vcovDeltaB(vcovB, vdelta)
}
\arguments{
\item{vcovB}{variance-covariance matrix of \eqn{vec(B)}.}

\item{vdelta}{numeric vector of covariate change.}
}
\value{
Squared numeric matrix of order two.
}
\description{
Given the covariance matrix of paramters, it return the covariance
matrix associated to the chanve in covariate values in the space
of covariantes.
}
\keyword{internal}
